+++ picomint-aleph-bft/src/dag/reconstruction/mod.rs
@@ -242,7 +242,7 @@ mod test {
     #[test]
     fn reconstructs_initial_units() {
         let mut reconstruction = Reconstruction::new();
-        for unit in &random_full_parent_units_up_to(0, NumPeers::new(4 as usize), 43)[0] {
+        for unit in &random_full_parent_units_up_to(0, NumPeers::new(4_usize), 43)[0] {
             let ReconstructionResult {
                 mut units,
                 requests,
@@ -258,7 +258,7 @@ mod test {
     #[test]
     fn reconstructs_units_coming_in_order() {
         let mut reconstruction = Reconstruction::new();
-        let dag = random_full_parent_units_up_to(7, NumPeers::new(4 as usize), 43);
+        let dag = random_full_parent_units_up_to(7, NumPeers::new(4_usize), 43);
         for units in &dag {
             for unit in units {
                 let round = unit.round();
@@ -293,7 +293,7 @@ mod test {
     #[test]
     fn requests_all_parents() {
         let mut reconstruction = Reconstruction::new();
-        let dag = random_full_parent_units_up_to(1, NumPeers::new(4 as usize), 43);
+        let dag = random_full_parent_units_up_to(1, NumPeers::new(4_usize), 43);
         let unit = dag
             .get(1)
             .expect("just created")
@@ -307,7 +307,7 @@ mod test {
     #[test]
     fn requests_single_parent() {
         let mut reconstruction = Reconstruction::new();
-        let dag = random_full_parent_units_up_to(1, NumPeers::new(4 as usize), 43);
+        let dag = random_full_parent_units_up_to(1, NumPeers::new(4_usize), 43);
         for unit in dag.first().expect("just created").iter().skip(1) {
             reconstruction.add_unit(unit.clone());
         }
@@ -321,14 +321,14 @@ mod test {
         assert_eq!(requests.len(), 1);
         assert_eq!(
             requests.last().expect("just checked"),
-            &Request::Coord(UnitCoord::new(0, PeerId::new(0 as u8)))
+            &Request::Coord(UnitCoord::new(0, PeerId::new(0_u8)))
         );
     }
 
     #[test]
     fn reconstructs_units_coming_in_reverse_order() {
         let mut reconstruction = Reconstruction::new();
-        let mut dag = random_full_parent_units_up_to(7, NumPeers::new(4 as usize), 43);
+        let mut dag = random_full_parent_units_up_to(7, NumPeers::new(4_usize), 43);
         dag.reverse();
         for units in dag.iter().take(7) {
             for unit in units {
@@ -350,7 +350,7 @@ mod test {
 
     #[test]
     fn handles_bad_hash() {
-        let node_count = NumPeers::new(7 as usize);
+        let node_count = NumPeers::new(7_usize);
         let mut reconstruction = Reconstruction::new();
         let dag = random_full_parent_units_up_to(0, node_count, 43);
         for unit in dag.first().expect("just created") {
@@ -402,7 +402,7 @@ mod test {
         const MAX_ROUND: Round = 7;
 
         let mut rng = rand::thread_rng();
-        let node_count = NumPeers::new(7 as usize);
+        let node_count = NumPeers::new(7_usize);
         let mut reconstruction = Reconstruction::new();
 
         let dag = random_full_parent_units_up_to(MAX_ROUND, node_count, 43);
